refactor(categories): add navigation to product details
