+++ lib/core/utils/shared_models/product_entity.dart
