+++ lib/features/categories/data/repo_impl/categories_repo_impl.dart
@@ -1,8 +1,8 @@
 import 'package:ecommerce_flower_app/core/utils/datasource_excution/api_manager.dart';
 import 'package:ecommerce_flower_app/core/utils/datasource_excution/api_result.dart';
+import 'package:ecommerce_flower_app/core/utils/shared_models/product_entity.dart';
 import 'package:ecommerce_flower_app/features/categories/data/data_source/contract/categories_remote_data_source.dart';
 import 'package:ecommerce_flower_app/features/categories/domain/entities/category_entity.dart';
-import 'package:ecommerce_flower_app/features/categories/domain/entities/product_entity.dart';
 import 'package:ecommerce_flower_app/features/categories/domain/repo/categories_repo.dart';
 import 'package:injectable/injectable.dart';
