--- lib/features/categories/domain/entities/product_entity.dart
@@ -1,27 +1,0 @@
-class ProductEntity {
-  final String? id;
-  final String? title;
-  final String? slug;
-  final String? description;
-  final String? imgCover;
-  final List<String>? images;
-  final int? price;
-  final int? priceAfterDiscount;
-  final int? quantity;
-  final int? discount;
-  final int? sold;
-
-  ProductEntity({
-    this.id,
-    this.title,
-    this.slug,
-    this.description,
-    this.imgCover,
-    this.images,
-    this.price,
-    this.priceAfterDiscount,
-    this.quantity,
-    this.discount,
-    this.sold,
-  });
-}
