+++ lib/features/categories/domain/use_cases/get_products_use_case.dart
@@ -1,5 +1,5 @@
 import 'package:ecommerce_flower_app/core/utils/datasource_excution/api_result.dart';
-import 'package:ecommerce_flower_app/features/categories/domain/entities/product_entity.dart';
+import 'package:ecommerce_flower_app/core/utils/shared_models/product_entity.dart';
 import 'package:ecommerce_flower_app/features/categories/domain/repo/categories_repo.dart';
 import 'package:injectable/injectable.dart';
