+++ lib/features/categories/presentation/view/widgets/product_item.dart
@@ -1,9 +1,10 @@
+
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:ecommerce_flower_app/core/assets/app_colors.dart';
 import 'package:ecommerce_flower_app/core/assets/app_icons.dart';
 import 'package:ecommerce_flower_app/core/utils/l10n/locale_keys.g.dart';
-import 'package:ecommerce_flower_app/features/categories/domain/entities/product_entity.dart';
+import 'package:ecommerce_flower_app/core/utils/shared_models/product_entity.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 
@@ -85,6 +86,9 @@ class ProductItem extends StatelessWidget {
             const SizedBox(height: 8),
             ElevatedButton.icon(
               onPressed: () {},
+              style: ElevatedButton.styleFrom(
+                padding: const EdgeInsets.symmetric(vertical: 6),
+              ),
               label: Text(
                 LocaleKeys.AddToCart.tr(),
                 style: Theme.of(context).textTheme.labelLarge!.copyWith(
