+++ lib/features/categories/presentation/view_model/categories_cubit.dart
@@ -2,8 +2,8 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:ecommerce_flower_app/core/base/base_state.dart';
 import 'package:ecommerce_flower_app/core/utils/datasource_excution/api_result.dart';
 import 'package:ecommerce_flower_app/core/utils/l10n/locale_keys.g.dart';
+import 'package:ecommerce_flower_app/core/utils/shared_models/product_entity.dart';
 import 'package:ecommerce_flower_app/features/categories/domain/entities/category_entity.dart';
-import 'package:ecommerce_flower_app/features/categories/domain/entities/product_entity.dart';
 import 'package:ecommerce_flower_app/features/categories/domain/use_cases/get_categories_use_case.dart';
 import 'package:ecommerce_flower_app/features/categories/domain/use_cases/get_products_use_case.dart';
 import 'package:ecommerce_flower_app/features/categories/presentation/view_model/categories_state.dart';
