+++ lib/features/product_details/presentation/view/widgets/image_view.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 import '../../../../../core/assets/app_colors.dart';
-import '../../../../home/domain/entity/product_entity.dart';
+import '../../../../../core/utils/shared_models/product_entity.dart';
 
 class ImageView extends StatefulWidget {
   final ProductEntity product;
