+++ lib/features/product_details/presentation/view/widgets/product_details_body.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 import '../../../../../core/assets/app_colors.dart';
-import '../../../../home/domain/entity/product_entity.dart';
+import '../../../../../core/utils/shared_models/product_entity.dart';
 
 class ProductDetailsBody extends StatelessWidget {
   final ProductEntity product;
